feat: add get product by id handler

+++ storefront/internal/entity/product.go
@@ -5,6 +5,6 @@ import "github.com/google/uuid"
 type Product struct {
 	UUID        uuid.UUID `json:"uuid"`
 	Name        string    `json:"name"`
-	Price       float64   `json:"price"`
+	Price       uint64    `json:"price"`
 	Description string    `json:"description"`
 }
